pyproject.toml
[project]
name = "news-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[tools]>=1.9.0",
    "langchain-google-genai>=4.2.0",
    "python-dotenv>=1.1.1",
    "supabase>=2.27.2",
    "tavily-python>=0.7.20",
]
